improve rviz visualization

+++ sv/util/math.cpp
@@ -4,7 +4,7 @@
 
 namespace sv {
 
-void MakeRightHanded(Eigen::Vector3f& eigvals, Eigen::Matrix3f& eigvecs) {
+void MakeRightHanded(Eigen::Vector3d& eigvals, Eigen::Matrix3d& eigvecs) {
   auto hand = eigvecs.col(0).cross(eigvecs.col(1)).dot(eigvecs.col(2));
   if (hand < 0) {
     eigvecs.col(0).swap(eigvecs.col(1));
